requirements.txt
customtkinter==5.2.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
watchdog==3.0.0
